+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Changelog
 
+### 1.2.0 (2021-06-21)
+
+#### Breaking changes
+  * None
+
+#### Enhancements
+  * Added sendChars method to the FlaNiumDriver class.
+ 
+#### Bug fixes
+  * Solving the problem with the impossibility of setting the FlaNiumDriverService port other than 9999
+
+
 ### 1.1.0 (2021-04-13)
 
 #### Breaking changes